requirements.txt
Django
django-jet
pytz
python-dotenv
django-filter
djangorestframework
bcrypt
